ncert-doubt-solver/backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
chromadb
pymupdf
pdfplumber
pytesseract
langdetect
python-multipart
pydantic
requests
python-dotenv
